Demo apps: State of the Union addresses

Two demo apps exploring the United States Presidential State of the Union addresses. The data are provided by the sotu package, and include all addresses through 2016. Interrupt R to stop the application (usually by pressing Ctrl+C or Esc).

run_sotu_app(...)

create_sotu_app()

run_sotu_decade_app(...)

create_sotu_decade_app()

Arguments

...
Arguments passed to explore()

Value

The run_sotu_* functions launch a Shiny app. The create_sotu_* functions return a corporaexplorerobject.

Details

For details, see https://kgjerde.github.io/corporaexplorer/articles/sotu.html.