Launching Shiny app for exploration of text collection

Launch Shiny app for exploration of text collection. Interrupt R to stop the application (usually by pressing Ctrl+C or Esc).

run_corpus_explorer(corpus_object, use_matrix = TRUE,
  regex_engine = c("default", "stringr", "re2r"),
  optional_info = FALSE, allow_unreasonable_patterns = FALSE,
  ui_options = list(), ...)

Arguments

corpus_object
A corporaexplorer object created by prepare_data.

use_matrix
Logical. If the corporaexplorerobject contains a document term matrix, should it be used for searches? (See prepare_data.)

regex_engine
Specify regular expression engine to be used:

`default`: use the re2r package (https://github.com/qinwf/re2r) for simple searches and the stringr package (https://github.com/tidyverse/stringr for complex regexes (i.e. when special regex characters are used).

`stringr`: use stringr for all searches.

`re2r`: use re2r for all searches.

optional_info
Logical. If TRUE, information about search method (regex engine and whether the search was conducted in the document term matrix or in the full text documents).

allow_unreasonable_patterns
Logical. If FALSE, the app will not allow patterns that will result in an enormous amount of hits or will lead to a very slow search. (Examples of such patterns will include '.' and '\b'.)

ui_options
List. Specify custom app settings. Currently available:

MAX_DOCS_IN_WALL_VIEW. Integer specifying the maximum number of documents to be rendered in the 'document wall' view. Default value is 12000.

font_size. Character string specifying font size in document viww, e.g. "10px"

...
Other arguments passed to runApp in the Shiny package.

Examples

# Constructing test data frame:
dates <- as.Date(paste(2011:2020, 1:10, 21:30, sep = "-"))
texts <- paste0(
  "This is a document about ", month.name[1:10], ". ",
  "This is not a document about ", rev(month.name[1:10]), "."
)
titles <- paste("Text", 1:10)
test_df <- tibble::tibble(Date = dates, Text = texts, Title = titles)

# Converting to corporaexplorerobject:
corpus <- prepare_data(test_df, corpus_name = "Test corpus")

#> Starting.

#> Document data frame done.

#> Calendar data frame done.

#> Document term matrix: text processed.

#> Document term matrix: tokenising completed.

#> Document term matrix: word list created.

#> Document term matrix done.

#> Done.

if(interactive()){
# Running exploration app:
run_corpus_explorer(corpus,
                    ui_options = list(MAX_DOCS_IN_WALL_VIEW = 12001))

# Running app to extract documents:
run_document_extractor(corpus)
}